fix: move isNativeDOMNode and other private utilities to their own files so they aren't included in the public API
