+++ src/utils/index.js
@@ -1,64 +1,4 @@
-/**
-	 * Abosolutizes a relative URL
-	 * @method _absolutizeURI
-	 * @private
-	 * @param {e.g., location.href} base
-	 * @param {URL to absolutize} href
-	 * @return {string} URL
-	 * Source: https://gist.github.com/Yaffle/1088850
-	 * Tests: http://skew.org/uri/uri_tests.html
-	 */
- var _absolutizeURI = function(base, href) {// RFC 3986
-
-	function removeDotSegments(input) {
-		var output = [];
-		input.replace(/^(\.\.?(\/|$))+/, '')
-			.replace(/\/(\.(\/|$))+/g, '/')
-			.replace(/\/\.\.$/, '/../')
-			.replace(/\/?[^\/]*/g, function (p) {
-				if (p === '/..') {
-					output.pop();
-				} else {
-					output.push(p);
-				}
-			});
-		return output.join('').replace(/^\//, input.charAt(0) === '/' ? '/' : '');
-	}
-
-	href = _parseURI(href || '');
-	base = _parseURI(base || '');
-
-	return !href || !base ? null : (href.protocol || base.protocol) +
-		(href.protocol || href.authority ? href.authority : base.authority) +
-		removeDotSegments(href.protocol || href.authority || href.pathname.charAt(0) === '/' ? href.pathname : (href.pathname ? ((base.authority && !base.pathname ? '/' : '') + base.pathname.slice(0, base.pathname.lastIndexOf('/') + 1) + href.pathname) : base.pathname)) +
-		(href.protocol || href.authority || href.pathname ? href.search : (href.search || base.search)) +
-		href.hash;
-};
-
-/**
- * Parses URI
- * @method _parseURI
- * @private
- * @param {The URL to parse} url
- * @return {Parsed URL} string
- * Source: https://gist.github.com/Yaffle/1088850
- * Tests: http://skew.org/uri/uri_tests.html
- */
-var _parseURI = function(url) {
-	var m = String(url).replace(/^\s+|\s+$/g, '').match(/^([^:\/?#]+:)?(\/\/(?:[^:@]*(?::[^:@]*)?@)?(([^:\/?#]*)(?::(\d*))?))?([^?#]*)(\?[^#]*)?(#[\s\S]*)?/);
-	// authority = '//' + user + ':' + pass '@' + hostname + ':' port
-	return (m ? {
-			href     : m[0] || '',
-			protocol : m[1] || '',
-			authority: m[2] || '',
-			host     : m[3] || '',
-			hostname : m[4] || '',
-			port     : m[5] || '',
-			pathname : m[6] || '',
-			search   : m[7] || '',
-			hash     : m[8] || ''
-		} : null);
-};
+import uri from './uri';
 
 export default {
 	/**
@@ -127,67 +67,11 @@ export default {
 	},
 	/**
 	 * Tests a URL to see if it's on the same domain (local) or not
-	 * @method isLocalRequest
+	 * @method isLocal
 	 * @param {URL to test} url
 	 * @return {bool} Whether the URL is local or not
-	 * Derived from: https://github.com/jquery/jquery/blob/master/src/ajax.js
 	 */
-	isLocalRequest: function(url){
-		var rurl = /^([\w.+-]+:)(?:\/\/([^\/?#:]*)(?::(\d+)|)|)/,
-			urlLower = url.toLowerCase(),
-			parts = rurl.exec( urlLower ),
-			ajaxLocation,
-			ajaxLocParts;
-
-		try {
-			ajaxLocation = location.href;
-		} catch( e ) {
-			// Use the href attribute of an A element
-			// since IE will modify it given document.location
-			ajaxLocation = document.createElement('a');
-			ajaxLocation.href = '';
-			ajaxLocation = ajaxLocation.href;
-		}
-
-		ajaxLocation = ajaxLocation.toLowerCase();
-
-		// uh oh, the url must be relative
-		// make it fully qualified and re-regex url
-		if (!parts){
-			urlLower = _absolutizeURI(ajaxLocation,urlLower).toLowerCase();
-			parts = rurl.exec( urlLower );
-		}
-
-		// Segment location into parts
-		ajaxLocParts = rurl.exec( ajaxLocation ) || [];
-
-		// do hostname and protocol and port of manifest URL match location.href? (a "local" request on the same domain)
-		var matched = !(parts &&
-				(parts[ 1 ] !== ajaxLocParts[ 1 ] || parts[ 2 ] !== ajaxLocParts[ 2 ] ||
-					(parts[ 3 ] || (parts[ 1 ] === 'http:' ? '80' : '443')) !==
-						(ajaxLocParts[ 3 ] || (ajaxLocParts[ 1 ] === 'http:' ? '80' : '443'))));
-
-		return matched;
-	},
-	/**
-	 * Utility method to determine whether or not the argument passed in is or is not a native dom node.
-	 * @method isNativeDOMNode
-	 * @param {object} testObject The object you want to check as native dom node.
-	 * @return {bool} Returns true if the object passed is a native dom node.
-	 */
-	isNativeDOMNode: function(testObject) {
-		var bIsNode = (
-			typeof Node === 'object' ? testObject instanceof Node :
-			testObject && typeof testObject === 'object' && typeof testObject.nodeType === 'number' && typeof testObject.nodeName === 'string'
-		);
-
-		var bIsElement = (
-			typeof HTMLElement === 'object' ? testObject instanceof HTMLElement : //DOM2
-			testObject && typeof testObject === 'object' && testObject.nodeType === 1 && typeof testObject.nodeName === 'string'
-		);
-
-		return (bIsNode || bIsElement);
-	},
+	isLocalRequest: uri.isLocal,
 	/**
 	 * A utility logging function to write messages or objects to the browser console. This is a proxy for the [`console` API](https://developers.google.com/chrome-developer-tools/docs/console).
 	 * @method log
